requirements.txt
flask==1.1.2
flask-wtf==0.14.3
email-validator==1.1.3
wtforms==2.3.3
flask-sqlalchemy==2.5.1
python-dotenv==0.18.0